Task-7 [Animation][New]: Implement item animation * Added repository which persist and provide base currency * Updated unit tests and added new ones * Fixed minor bugs * Updated demo version

+++ repositories/remote/src/main/java/io/mellouk/repositories/remote/dto/RateDto.kt
@@ -2,7 +2,7 @@ package io.mellouk.repositories.remote.dto
 
 data class RateDto(
     val currency: CurrencyDto,
-    val value: Double
+    val value: Float
 )
 
 enum class CurrencyDto(val code: String) {
